moved sdk files to sdk folder
